Update Author.jsx
I think it is better to use flex and justify content center so as to center the image as the author image is not centering even with the code here on github. Also, the next js Image tag doesn't accept width and height attributes in "" so width = {100} instead of width= "100px". Thanks for this beautiful project. May God continue to empower you